requirements.txt
streamlit
pandas
inventree
python-dotenv
# Add pytest later when tests are implemented
# Add black later for formatting checks if needed in CI/CD